first ui library component implemented & tweaks to form on index
